Transpile all non-external js files with babel
